Revert "Prioritize alternate charset for SSID translation"
Revert submission 23579364

Reason for revert: It seems more likely that actual UTF-8 SSIDs will be misinterpreted as GBK, than GBK being misinterpreted as UTF-8. This is because GBK (and the other common extended ASCII charsets like EUC-KR) has valid characters mapped to nearly every 2-byte sequence, which means any UTF-8 string with an even number of bytes has a high likelyhood of being interpreted as GBK.

The use of the "is UTF-8" bit is also not guaranteed (and seems to be uncommon), so the lack of it is not a good indication that the SSID is not UTF-8. Thus to be on the safe side, we should revert back to preferring UTF-8 over the alternate charset (which is also consistent with the OEM GBK solutions). 
 
Reverted changes: /q/submissionid:23579364

Change-Id: Idde53da443739724cb32bd23ff637fadadc9723c