Fix supplicant SSIDs being interpreted with the wrong encoding
By default, SSID bytes that are both valid UTF-8 and an alternate
encoding like GBK will be interpreted as UTF-8. However, this means
ambiguous SSIDs reported by supplicant will be interpreted as UTF-8 even
if we connected to a GBK AP, causing a mismatch in the framework.

To avoid this mismatch, record which SSID translation is intended when
setting the supplicant SSID, and return that translation instead of the
default translation when the supplicant sends SSIDs back to the framework.

Bug: 283216966
Test: atest FrameworksWifiTests
Change-Id: I675beb941afe33783e6a625a35f3f177a1b8cee7