Set hasEverConnected to true for Passpoint configs with an SSID.
SSID is only set if the Passpoint network has
connected since the last reboot.

This fixes an issue when networks are selected for
PNO scans. Without this fix, all Passpoint networks
retrieved from the PasspointManager
are non-ephemeral and have hasEverConnected = false.
This causes them to get filtered out of the PNO
SSID list.

Bug: 195659300
Test: atest PasspointManagerTest WifiConnectivityManagerTest
Test: Manual test
       1. Connect to a Passpoint network.
       2. Connect to a hotspot.
       3. Turn off the screen.
       4. Disable the hotspot.
       5. Device should trigger a PNO scan and
          connnect to the Passpoint network.
Change-Id: I6d17682c505003621c89f8765ecb88f860d77f88