Implement AFCLocation and AFCLocationUtil classes
Provide a framework to convert locations to coarse location objects needed for AFC queries and determine if points are inside of a coarse location.

Test: Verified that an ellipse is created with the correct randomized
fields within a range given positive coordinates.
Test: Verified that an ellipse is created with the correct randomized
fields within a range given negative coordinates.
Test: Verified that a longitude point that moves gradually from the
center to outside an AFCLocation will be detected accordingly.
Test: Verified that a latitude point that moves gradually from the
center to outside an AFCLocation will be detected accordingly.
Test: atest AFCLocationTest
Test: atest AFCLocationUtilTest

Bug: 288195822
Change-Id: Ica5fb157d5ebc7e759ed98e6c8021ad4f84717eb