Notify CC change even if the new CC is the same as the old
In Japan, 5GHz channels are either indoor or DFS and are not allowed for
SAP by default. However, if STA is connected via a 5GHz channel, then
we may use that channel for SAP, and the driver will notify the change
of available channels via a CC change event. But since the framework
ignores CC change events if the new CC is the same as the old one, we
should remove this check and still notify CC listeners in this case.

Bug: 287936895
Test: atest FrameworksWifiTests
Change-Id: I8af435a69b3d51341ff3508537c9c646a06dd89d