Notify CC change even if the new CC is the same as the old
In Japan, 5GHz channels are either indoor or DFS and are not allowed for
SAP by default. However, if STA is connected via a 5GHz channel, then
we may use that channel for SAP, and the driver will notify the change
of available channels via a CC change event. But since the framework
ignores CC change events if the new CC is the same as the old one, we
should remove this check and still notify CC listeners in this case.

Bug: 287936895
Test: atest FrameworksWifiTests
Change-Id: I378659819237452dd8796cac4ac6c36a1cc9ea85
Merged-In: I378659819237452dd8796cac4ac6c36a1cc9ea85
(cherry picked from commit 39349c5b29bba3cf1243444e66fc296536e549c7)