Implement AfcClient class
Create AfcClient framework responsible for making HTTP post request queries to the AFC server and receiving back the AfcServerResponse object.

Implement a shell command that allows the caller to configure the AFC
server URL and request properties to create HTTP requests. It is called
with the "configure-afc-server" argument. Calling this configures the
server and won’t cause additional side effects on the device.

Test: Verified that the HTTP request is generated with expected fields and that the request ID is incremented by 1 with each query to the AFC server.
Test: Verified that the response received from the HttpClient request matches that of the response from the AfcClient.
Test: atest AfcClientTest

Bug: 288195475
Change-Id: I6a211a72b1fb8431fa90d5c239de7fb9ba670f46