Replace SSID with network ID check in TOFU_CERTIFICATE_EVENT handler
The previous connecting TOFU wifi configuration is
cached in InsecureEapNetworkHandler. When TOFU_CERTIFICATE_EVENT
is handled in InsecureEapNetworkHandler, process the event only
if current connecting configuration network ID is same as cached
TOFU configuration network ID. This is done to avoid processing
the event for a NON-TOFU network with same SSID as cached TOFU
config.

Bug: 289745107
Test: atest ClientModeImplTest InsecureEapNetworkHandlerTest
Test: Manual - Connect/Disconnect using TOFU
Change-Id: Iea1d52eb66077a27d0443da8d3924c7ef95d1e23