Check whether shell command was used to trigger an AFC query
Edit the configure-afc-server wifi shell command to cache the server URL and request properties in the AfcManager, and edit the trigger-afc-location-update shell command to inform the AfcManager that the location update was triggered from a shell command. If a location update was triggered from the shell command, then set the server URL and request properties to the cached values, otherwise, continue with normal querying logic of getting the server URL from the overlay.

Test: Manually tested configure-afc-server shell command to verify that values are correctly cached.
Test: Manually tested trigger-afc-location-update shell command to verify that the cached URL and request properties are being used in a query.
Test: Verified that a location change initiated from a shell command will override whether AFC is supported for that device.
Test: atest AfcManagerTest

Bug: 290073935
Change-Id: I9f06cb20cc34dbcf6511208f8a4400d58461cbe5