call p2p listen to wait for the invitation from peer
As per P2P specification, 3.1.5.3 "A P2P Device that receives an Invitation Request frame
to re-invoke a persistent Group and responds with a P2P Invitation Response frame with a
Status attribute with the Status Code field set to “Fail: information is currently
unavailable” shall restart the Invitation procedure by sending an Invitation Request
frame to the requesting P2P Device". So the device which invited the peer should wait for
the new request from Peer. The supplicant implementation moves the P2P state to IDLE
after receiving the wait request. So the frames from the Peer device are
not received. This change is to call the p2p listen to start listening
for the frames. Also the listen/discovery process should be stopped
immediatly after the group formation.

Bug: 290876130
Test: Manual - Customer tested persistent P2P connection with Windows PC
Test: atest WifiP2pServiceImplTest

Change-Id: I065ae76702cd48e320bcac89b33f690c75d67aa8