Change Logging and AFC Response Code Handling
Update the framework so that logging is more descriptive and clear.

Change AfcServerResponse#parseAfcResponseCode so it returns 0 upon not finding an AFC response code because the internal Google AFC server doesn't provide an AFC response code upon a successful query. This ensures that successful queries are completely parsed and returned as successful instead of exiting early with an error.

Test: None
Bug: 287358930
Change-Id: Ia2abf0137be36416e6dd1bb162a01e27c2af56a5