Update Checking Location In Bounds Method
Update the AfcEllipseLocation#checkLocation method to convert the major and minor axes to degrees since the in bounds formula assumes the same units for the axes as the center coordinates of the location points. Converting the axes from meters to degrees is done by dividing each axis in meters by 111139.

Accounted for the edge case where the ellipse crosses the minimum or maximum longitude values to ensure that location points on the other side of the axis can be classified as in bounds if they are inside where the ellipse would cover. For example, if one point has longitude -179.9999° and latitude 37° and the other point has longitude 179.9999° and latitude 37°, then this will be classified as in bounds. Latitude values don't have this edge case because they don't wrap around the globe 360 degrees like longitude values do.

Test: Verified that an ellipse that crosses the maximum longitude accurately checks the bounds for locations.
Test: Verified that an ellipse that crosses the minimum longitude accurately checks the bounds for locations.
Bug: 288195822

Change-Id: I5df549e9da3925a3a18f3d47a243306bb7ad6230