Move client register for PNO scan to inside statemachine
registering the new client should be done in the statemachine since
unregistering it (stopPnoScan) is done inside the statemachine.

In the current code startPnoScan registers the client before sending the
statemachine command, which causes the following race:
- stopPnoScan post command to remove client
- startPnoScan registers the client and then post command to start scan
- CMD_STOP_PNO_SCAN removes the newly registered client
- CMD_START_PNO_SCAN fails to find a client and aborts

This results in PNO scan to silently fail when stopPnoScan and
startPnoScan are called back-to-back.

Bug: 294593727
Test: manually verified PNO scanning still works
Test: atest com.android.server.wifi
Change-Id: I14ec675c26a5827d0966f120900ba8a72115da80