[HIDL] Use the PIN from supplicant in case of P2P_PROV_DISC_ENTER_PIN_EVENT
In some implementations, the PIN is shared to supplicant via OOB
mechanisms. The intention is to connect without popup.
So if the PIN is available from Supplicant, set that PIN in framework
and connect without a dialog box to enter PIN.
This is a continuation of change id - I466f8c5e247a8c13f851f30ce8b23913aa3d2d4d
where HIDL implementation was not handled.

Bug: 282592275
Test: WiFi Direct CtsVerifier test
Change-Id: Ib4a95cfa9f63bb14bc8dc20c6ea740d7f95a8e4d