Use the security params set in connection for deriving security type
When the connection sets PMF required bit to true when using
authentication key management - WPA_EAP_SHA256, it is a
WPA3-Enterprise connection. So while deriving the security
type from AKM, use the security parameters configured in
the supplicant during connection.

Bug: 300352523
Test: atest WifiConfigurationTest
Test: Manual - Configure AP in WPA3 Enterprise mode.
On the phone, manually add the same network with security type set
to WPA/WP2-Enterprise. After the connection, confirmed from
the network details page that the security type is set to
WPA3-Enterprise.

Change-Id: Icc9a9d41fe16fe46cd8c9fce81ed1e878259b291