Place Passpoint networks at the beginning of the PNO network list.
The current network sort comparator is based on
LruConnectionTracker, which is never updated for
Passpoint networks. As a result, Passpoint networks
would always be moved to the end of the PNO network
list during sorting.

Instead of modifying the LruConnectionTracker, we can
reserve up to 2 slots at the head of the PNO network
list for recently-connected Passpoint networks.

Bug: 300069792
Test: atest WifiConnectivityManagerTest
Test: Manual test
       1. Connect to a Passpoint network.
       2. Connect to several non-Passpoint networks.
       3. Connect to a hotspot.
       4. Turn off the screen on the DUT and disable the
          hotspot to trigger a PNO scan.
       5. From the logs, verify that the PNO network list
          order is the Passpoint network, followed by the
	  non-Passpoint networks in MRU order.
Change-Id: I36e16e6c9c88a456c9d0f8069a7ce25f2dacecfc