Don't send BLOCK_DISCOVERY when P2P is in WaitingState
Before triggering the DHCP, clientModeImpl send BLOCK_DISCOVERY
to stop the P2P discovery. And wait for CMD_PRE_DHCP_ACTION_COMPLETE
message back from P2P module. This is done to protect the DHCP exchange.
But if p2p is in a waiting state for user response to create the interface,
all the messages will be deferred. So clientModeImpl doesn't get
CMD_PRE_DHCP_ACTION_COMPLETE response from P2P module. This blocks the DHCP
when p2p is in waiting state. So added changes to check if p2p is in waiting state
and if its in the waiting state, don't send BLOCK_DISCOVERY message.

Bug: 297808149
Test: atest ClientModeImplTest
Test: Manual - Tested STA+P2P concurrency scenario and make sure that
the waiting state is updated properly while creating the interface.

Change-Id: Iacf3e35e82c3cf06e9986dba3535a1ac71b4dd88