Correct evaluation of whether network selection is needed in multi-internet case
When secondary internet is being requested, a list of candidate will be
returned regardless of whether the current primary is sufficient or not.
So, to avoid unintentionally running network seletion on primary, we
need to explicitly do the sufficiency check evaluation on the primary
when secondary candiate is not found.

Bug: 297333966
Test: atest com.android.server.wifi
Change-Id: I2fdda90f4c889902d27226169c999995ecedf8ce