Restart SoftAp with original SoftApConfiguration bands on CC change
Currently, SoftAp is restarted on a CC change using the cached
mCurrentSoftApConfiguration. However, the bands of
mCurrentSoftApConfiguration are modified from the originally specified
config based on regulatory, interface concurrency, and STA status
restrictions. Thus a CC change to a more restrictive CC (like world
mode) and back will result in the AP bands retaining the restrictions
from world mode.

To fix this, restart AP with the original bands instead of the current
ones.

Bug: 285262593
Test: atest WifiCountryCodeTest SoftApManagerTest
Change-Id: If7998d9b21693b66958ee9385cbf708566e2397c