Fix ICM dialogs getting stuck if aware becomes opportunistic
ICM dialogs are bypassed for an opportunistic aware request. However,
the request may become opportunistic before an active dialog is replied
to, causing us to bypass the reply message. This causes ICM to be stuck
waiting for the user reply.

Instead of bypassing, we should always route the request to ICM, and
explicitly pass in the bypass flag so that ICM can correctly reset its
state.

Bug: 308865359
Test: atest WifiAwareStateManagerTest
Change-Id: Ib609ec0731c604aec8e96ac96fe175ddc3d01a32