Fix Auto LOHS not starting with desired bands when driver isn't up
Automotive can specify 5Ghz and 6Ghz bands for LOHS if the overlays for
them are enabled and there are available channels. However, if the
driver isn't up yet, then there will be no available channels in the
current SoftApCapability, causing us to default to 2.4Ghz.

Instead of checking the current available bands, we should always send
down a combined 2GHz + 5Ghz or 6Ghz band as long as the overlays allow
them, and then rely on the driver ACS to select the available channel.

Bug: 305071691
Test: atest WifiApConfigStoreTest
Change-Id: I7f8ae9dc0c87dd7243ba916a74a5ad1406d00560