Reduce the passpoint matching for each network selection
Matching passpoint config to the ScanResults is necessary for each time new ScanResults available for both settings display and network selection.
It was running one time per nominator with the same input. This change will make it running before the nominate starts to reduce the matching process to once per new ScanResults available.

Bug: 295398783
Test: atest com.android.server.wifi
Change-Id: Ic4ac6ef919a813e85a91b031e4871a48f1a5cf2e