Disable network permanently if no internet validation had ever passed
Enhance the blocklist behavior to permanently disable a network if
network validation have never passed. This solves a corner case
caused by the user disconnecting wifi when network validation is still
in-progress, causing a no-internet network to never get permanenly
blocked afterwards.

Networks that already exist on the device when this fix goes live will
be treated as if they successfully passed network validation previously.

Bug: 306113405
Test: atest com.android.server.wifi
Test: manually verified

Change-Id: I8f4f87ef594cc650c0e5208b097ab77ab013da98