Rely on S+ HDM interface priority for enabling Wifi when tethering
Enabling Wifi isn't allowed for non-privileged apps if tethering AP is
up. This was meant to prevent the non-privileged STAs from deleting
tethering AP in STA|AP devices, back when the iface priority scheme
was based on iface type alone. However, this results in apps not being
allowed to enable wifi even for STA+AP devices.

For S and above, we should remove this check and rely on the S+
worksource-based priority scheme, which already considers the privilege
 of the iface worksources. This will result in non-priveged STAs not
being able to delete privileged tethering AP in STA|AP devices, while
allowing the STA to be created for STA+AP devices.

Bug: 312264246
Test: atest WifiServiceImplTest

Change-Id: Idc165f54c24c7531c0273be92bfa0c42f73a4f74