Add RTT busy status and retry duration API
This change introduces a new status code, STATUS_BUSY_TRY_LATER, to the
RangingResult API. This allows a Wi-Fi RTT responder to indicate that it
is temporarily unable to handle a ranging request.

A corresponding method, getRetryAfterDurationMillis(), is added to
provide the requester with a suggested delay before retrying the
operation.

The retryAfterDuration field from the HAL RttResult (in seconds) is now
plumbed through to the framework and converted to milliseconds.

A CTS test is added to validate the functionality of the new status and
API method.

Bug: 407453160
Flag: com.android.wifi.flags.rtt_busy_try_later_api
Test: atest CtsWifiTestCases:android.net.wifi.rtt.cts.WifiRttTest#testRan
gingResultBusyTryLaterStatus
Low-Coverage-Reason: wrong detection? New code is covered by the new test-cases

Change-Id: I210df72bd88e6ea5e48c3ea42bb82d92d8cabd61