Fix NullPointerException in UsdServiceDiscovery
The `WifiP2pUsdBasedServiceDiscoveryConfig` and
`WifiP2pUsdBasedLocalServiceAdvertisementConfig`
can return null for `getServiceSpecificInfo()`
and `getFrequenciesMhz()`.
This can lead to `NullPointerException`
when these values are passed to the AIDL interface.

This change adds null checks for `serviceSpecificInfo`
and `frequenciesMhz` and provides empty arrays as
default values if they are null, preventing the `NullPointerException`.

Bug: 446623186
Flag: com.android.wifi.flags.wifi_direct_r2
Test: atest testDiscoverUsdBasedServices
Test: atest SupplicantP2pIfaceHalAidlVendorImplTest
Change-Id: I65b913523eca4a3152a3bb60c3166f01e0d08287