Avoid multiple dialogs for wifi network specifier
When the app is already provided with user disconnect reason, framework will trust the app to handle when to trigger reconnect.

So, don't temporarily disable the network, as that will cause another dialog to pop up when the app request connection.

Flag: com.android.wifi.flags.local_only_disconnect_reason
Bug: 449257685
Test: atest com.android.server.wifi
Change-Id: I29c2f5812c5afc7d8859d1ea05289eb354c63852