Add PD USD framework implementation
1) Translates isProximityRangingEnabled from SDK config to HAL config 2)
Add isProximityRangingEnabled to usdCapabilities and getting it from
   the UsdManager#getCharacteristics()
3) Translates publishType from SDK config to HAL config 4) Translates
selfDevIk and peerDevIks from SDK config to HAL config 5) Replaces the
serviceName in the subscribeConfig with null if the
   value equals "SERVICE_NAME_ANY"

Bug: 466265351
Test: m -j 32
Flag: com.android.wifi.flags.proximity_ranging
Low-Coverage-Reason: Bypass. Basic tests are added now. Detailed tests will be added in a separate CL along with implementation
Change-Id: Ibacb22873236e331744e3094cba911a717737a36